Move errors to an error file
